Python Projects/Crypto Pipeline and Tensorflow/requirements.txt
absl-py==2.3.1
astunparse==1.6.3
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
contourpy==1.3.3
cycler==0.12.1
Flask==3.1.2
flask-cors==6.0.1
flatbuffers==25.9.23
fonttools==4.60.1
gast==0.6.0
google-pasta==0.2.0
greenlet==3.2.4
grpcio==1.75.1
h5py==3.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
keras==3.11.3
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.6
mdurl==0.1.2
ml_dtypes==0.5.3
namex==0.1.0
numpy==2.3.3
nvidia-cublas-cu12==12.9.1.4
nvidia-cuda-cupti-cu12==12.9.79
nvidia-cuda-nvcc-cu12==12.9.86
nvidia-cuda-nvrtc-cu12==12.9.86
nvidia-cuda-runtime-cu12==12.9.79
nvidia-cudnn-cu12==9.13.1.26
nvidia-cufft-cu12==11.4.1.4
nvidia-curand-cu12==10.3.10.19
nvidia-cusolver-cu12==11.7.5.82
nvidia-cusparse-cu12==12.5.10.65
nvidia-nccl-cu12==2.28.3
nvidia-nvjitlink-cu12==12.9.86
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
protobuf==6.32.1
psycopg2-binary==2.9.10
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
rich==14.1.0
scikit-learn==1.7.2
scipy==1.16.2
setuptools==80.9.0
six==1.17.0
SQLAlchemy==2.0.43
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3
